abstract datagram code away
